[!!!][TASK] Add type declarations to ViewHelper classes
Type declarations are added to all ViewHelper methods as well as
all properties that are considered internal and are most likely not
redefined directly by a ViewHelper implementation. Three properties
are left out intentionally, since they are currently part of the
ViewHelper API:

* `AbstractViewHelper::$escapeOutput`
* `AbstractViewHelper::$escapeChildren`
* `AbstractTagBasedViewHelper::$tagName`

Adding type declarations to those properties would make it impossible
to create ViewHelpers that are compatible with Fluid v4 and v5.