[FEATURE] Introduce <f:enum.tryFrom> ViewHelper
A `<f:enum.tryFrom>` ViewHelper can be useful to map a given value to a backed
enum case to unlock the potential of the backed enums also in Fluid. This way,
not only the name and value of a backed enum can be retrieved, also arbitrary
methods defined in the backed enum can be called.