[FEATURE] Support union types for ViewHelper and component API (#1226)
With the `LenientArgumentProcessor`, which was previously used for
ViewHelper arguments, it was already possible to specify multiple
types for an argument. However, that definition was silently ignored
during validation (see `ImageViewHelper` in TYPO3 Core).

With the introduction of the `StrictArgumentProcessor` and its usage
for ViewHelpers in Fluid 5, this behavior changed: If a union type
is specified for an argument and a value with one of the allowed types
is supplied, this currently results in an exception because the
argument processor doesn't know about union types.

This patch officially adds support for union types in Fluid. It is
now possible to define more than one allowed type for an argument,
both in ViewHelpers and in components:

ViewHelper example:

```php
$this->registerArgument('arg', 'array|string', '');
```

Component example:

```xml
<f:argument name="arg" type="array|string" />
```

Note that automatic type casting cannot happen for union types because
it's not possible to determine the desired type automatically. Thus,
if union types are used, all possible types need to be specified
explicitly and need to be dealt manually in the ViewHelper's
implementation.

Resolves: #1130