[FEATURE] Overhaul template path resolving (#1243)
All path handling and resolving in Fluid is performed by the
`TemplatePaths` class, which is attached to the current rendering
context. While this class has already been simplified in the past,
the actual path resolving has largely stayed the same because it
includes edge cases that were moved over from TYPO3.

This patch attempts to make the path resolving code more understandable,
while also eliminating some shortcomings. Most importantly, it is no
longer required for template names to start with an uppercase character.
A fallback chain is introduced, which prefers the provided spelling
of the template name but tries its uppercase variant in case the
template cannot be found.

In a follow-up patch, this will also allow us to introduce a custom
file extension for Fluid templates, which will then also be part of that
fallback chain.

Resolves: #371
Related: #938