[TASK] Use Fluid file extension in examples
As a follow-up to #1258, all example templates are now using the new
Fluid file extension. In all places where the file extension isn't
specified explicitly, the fallback mechanism allows us to just rename
the templates without any other changes. However, for
`setTemplatePathAndFilename()`, an explicit path is given, which
needs to be adjusted. Also, if the file extension is specified
explicitly when rendering a partial, this also needs to be adjusted
manually.