[FEATURE] <f:contains> ViewHelper (#1193) (#1262)
This patch adds a new `<f:contains>` ViewHelper to Fluid, which allows
to check if a provided string or array contains a specified value.
Depending on the input, this mimicks PHP's `in_array()` or `str_contains()`.

The initial version of this patch is intended to be backported to Fluid 4.
In a follow-up, the code and argument types can be improved for Fluid 5.

Resolves: #1125

Co-authored-by: Michael Staatz <michael@staatzstreich.de>