[FEATURE] Overhauled cache warmup command (#1224)
Previous Fluid versions already had a CLI command which could be used
to warmup the cache for template files, e. g. during deployment.
However, the previous implementation required a lot of additional
context and mixed parsing/compilation and rendering state. The
previous warmup implementation has already been deprecated with
Fluid v4 and removed from Fluid v5 and is now replaced with a
more straightforward implementation.

The new CLI command needs to be supplied with a list of paths that
should be scanned for template files as well as the location of
the template cache. By default, the command uses the newly introduced
Fluid file extension (see #1258) to identify Fluid templates.
However, optionally an alternative file extension can be supplied
as well.

```
bin/fluid warmup \
    --cacheDirectory examples/cache/ \
    --path examples/Resources/

bin/fluid warmup \
    --cacheDirectory examples/cache/ \
    --path examples/Resources/ \
    --path examples/ResourceOverrides/

bin/fluid warmup \
    --cacheDirectory examples/cache/ \
    --path examples/Resources/ \
    --extension html
```

In addition to the cache warmup, the command collects and outputs
exceptions and deprecations that happened during parsing and
compilation of the found template files. The underlying API,
namely `TemplateFinder` for finding template files and
`TemplateValidator`, which produces a validation report for each
template, are marked as `@internal` for now.

Resolves: #1103