[BUGFIX] Ensure valid name for cache classes (#1268)
Due to refactorings in the cache layer of Fluid, which were necessary
to make cache warmup feasible, the cache identifiers have changed
with Fluid 5. This currently leads to parse errors when a template
name starts with a number because the associated cache file and most
importantly the generated PHP class name also starts with that number,
which isn't allowed in PHP.

This patch addresses that issue by prepending "template_" to cache
identifiers. Other special characters are already dealt with in
`TemplateCompiler`, so we only need to ensure that the identifier
starts with an alpha-character.

Resolves: #1266