[BUGFIX] Allow constants and enum cases as argument default
Due to the way the structure of components is extracted from template
files, it was previously not possible to use a PHP constant or enum
case as default value for a component argument.

This patch adds the `<f:constant>` ViewHelper to the list of allowed
ViewHelpers for that pre-parsing step, which makes it now possible
to use `<f:constant>` from within `<f:argument>`.