[TASK] Allow TemplateValidator to discover underscore variables (#1283)
Fluid 5 disallows variable names that start with an underscore. While
this is already enforced in `StandardVariableProvider` for setting
such variables, this change now also triggers an exception if a
variable with an invalid name is accessed.

This is probably a temporary change to be removed with Fluid 6. For now,
it allows the `TemplateValidator`, which is used during template
warmup, to discover this issue during parse time, and to let users
know about affected templates.

Resolves: #1263